+++ Chasm.Compatibility/Chasm.Compatibility.Unsafe/Unsafe.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable CommentTypo GrammarMistakeInComment
 
-#if (NETSTANDARD1_0_OR_GREATER || NET45_OR_GREATER || NETCOREAPP1_0_OR_GREATER)
+#if NETCOREAPP1_0_OR_GREATER || NETSTANDARD1_0_OR_GREATER || NET45_OR_GREATER
 using System.Runtime.CompilerServices;
 [assembly: TypeForwardedTo(typeof(Unsafe))]
 #else